# make ImageItem combine levels + lut not be a pessimization