This PR continues #1630 in trying to fix the pessimization identified in #1590

Cases tested with PYQTGRAPHPROFILE=functions.makeARGB:
1) VideoSpeedTest.py --size=3072x3072 --levels=0,255
   * both levels and lut bypassed (fastest path)
2) VideoSpeedTest.py --size=3072x3072 --levels=50,200
   * levels applied, lut bypassed
3) VideoSpeedTest.py --size=3072x3072 --lut
   * levels bypassed, lut applied (was previously both applied)
4) VideoSpeedTest.py --size=3072x3072 --levels=50,200 --lut
   * levels bypassed, lut applied (was previously both applied)
5) VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --levels=0,65535
   * levels applied, lut bypassed
6) VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --levels=100,10000
   * levels applied, lut bypassed
7) VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --lut
   * levels bypassed, lut applied (was previously both applied)
8) VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --levels=100,10000 --lut
   * levels bypassed, lut applied (was previously both applied)
 
#1630 fixed cases 1,2,5,6
This PR fixes cases 3,4,7,8

It would good to take a look at #792 and #793 to get context on what this PR is working around.
What this PR does is to explicitly supply a ```scale``` argument with a specific value to ```makeARGB()``` such that ```makeARGB()``` does not use its own default. This causes rescaling to be bypassed, which was the intention of combining levels + lut in the first place.