@pijyoi 2021-03-28
I have tried to restore the behavior to pre-#793, that is to say, levels-only is converted to lut-only. i.e. ```rescaleData()``` is traded away for ```applyLookupTable()```

VideoSpeedTest.py --size=3072x3072
both levels and lut bypassed (fastest path)
VideoSpeedTest.py --size=3072x3072 --levels=0,255
both levels and lut bypassed (equivalent to not specifying levels)

all the following cases bypass levels and applies lut only
VideoSpeedTest.py --size=3072x3072 --levels=50,200
VideoSpeedTest.py --size=3072x3072 --lut
VideoSpeedTest.py --size=3072x3072 --levels=50,200 --lut
VideoSpeedTest.py --dtype=uint16 --size=3072x3072
VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --levels=0,65535
VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --levels=100,10000
VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --lut
VideoSpeedTest.py --dtype=uint16 --size=3072x3072 --levels=100,10000 --lut

This would also allow us to drop the fits_int32 specialization in #1648.