@pijyoi 2021-03-28
Timings on a Linux machine with numpy 1.20.2
Results suggest that this PR commit 2 converting rescale-only to lut-only is a pessimization.
So if we are not seeking to restore behavior to pre-#793, then this PR commit 1 gives the better outcome.

python examples/VideoSpeedTest.py --size=3072x3072 --levels=50,200
```
master
> Entering functions.makeARGB
  check inputs: 0.0100 ms
  apply levels: 20.3686 ms
  apply lut: 0.0098 ms
  allocate: 0.0012 ms
  reorder channels: 5.0273 ms
  alpha channel: 0.0057 ms
< Exiting functions.makeARGB, total time: 25.4271 ms

this PR commit 1
> Entering functions.makeARGB
  check inputs: 0.0134 ms
  apply levels: 19.1901 ms
  apply lut: 0.0083 ms
  allocate: 0.0014 ms
  reorder channels: 5.2166 ms
  alpha channel: 0.0050 ms
< Exiting functions.makeARGB, total time: 24.4391 ms

this PR commit 2
> Entering functions.makeARGB
  check inputs: 0.0119 ms
  apply levels: 0.0083 ms
  apply lut: 39.0100 ms
  allocate: 0.0067 ms
  reorder channels: 6.9034 ms
  alpha channel: 0.0062 ms
< Exiting functions.makeARGB, total time: 45.9511 ms
```

python examples/VideoSpeedTest.py --size=3072x3072 --levels=50,200 --lut
```
master
> Entering functions.makeARGB
  check inputs: 0.0129 ms
  apply levels: 19.0156 ms
  apply lut: 38.1424 ms
  allocate: 0.0057 ms
  reorder channels: 5.4858 ms
  alpha channel: 0.0060 ms
< Exiting functions.makeARGB, total time: 62.6724 ms

this PR commit 1
> Entering functions.makeARGB
  check inputs: 0.0145 ms
  apply levels: 0.0086 ms
  apply lut: 40.4389 ms
  allocate: 0.0062 ms
  reorder channels: 6.0117 ms
  alpha channel: 0.0062 ms
< Exiting functions.makeARGB, total time: 46.4907 ms

this PR commit 2
> Entering functions.makeARGB
  check inputs: 0.0114 ms
  apply levels: 0.0081 ms
  apply lut: 38.5458 ms
  allocate: 0.0062 ms
  reorder channels: 5.5625 ms
  alpha channel: 0.0062 ms
< Exiting functions.makeARGB, total time: 44.1446 ms
```