@j9ac9k 2021-03-28
> So if we are not seeking to restore behavior to pre-#793, then this PR commit 1 gives the better outcome.

I'm going to have to defer to @outofculture and @campagnola for input here.  Right now we have our hands full with writing a submission for a conference; so it might be a few days before you get helpful follow up.

My stance regarding #793 behavior is that ideally the issue presented in #792 (where `scale=3` ideally remains resolved)