@pijyoi 2021-03-29
>I'm going to have to defer to @outofculture and @campagnola for input here. Right now we have our hands full with writing a submission for a conference; so it might be a few days before you get helpful follow up.

No worries.

>My stance regarding #793 behavior is that ideally the issue presented in #792 (where scale=3 ideally remains resolved)

I made the first 2 commits erroneously thinking that issue #792 didn't apply to integer data.
I think I have fixed that in the last commit. During levels + lut combination, the scale value used needs to include the fix done in #793.

It looks a bit kludgy, the part that calls makeARGB() with a specially chosen scale value in order to cause apply levels to be bypassed.